+++ .eslintrc.json
@@ -4,7 +4,7 @@
   "extends": ["eslint:recommended"],
   "overrides": [
     {
-      "files": ["**/*.ts"],
+      "files": ["**/*.mts"],
       "extends": ["plugin:@typescript-eslint/recommended"]
     }
   ]
