+++ action.yml
@@ -32,4 +32,4 @@ inputs:
     required: false
 runs:
   using: node16
-  main: dist/index.js
+  main: dist/index.mjs
